Moved UI related stuff to one package
