+++ src/main/kotlin/brzzzn/fabadditions/item/warp/AbstractWarpStaffItem.kt
@@ -7,12 +7,12 @@ import brzzzn.fabadditions.data.WarpPosition
 import brzzzn.fabadditions.framework.Vector2
 import brzzzn.fabadditions.framework.Vector3
 import brzzzn.fabadditions.framework.WorldRef
-import brzzzn.fabadditions.guis.warpstaff.WarpGui
+import brzzzn.fabadditions.ui.screens.guis.warpstaff.WarpGui
 import brzzzn.fabadditions.item.warp.network.AddPosition
 import brzzzn.fabadditions.item.warp.network.DeleteWarp
 import brzzzn.fabadditions.item.warp.network.OpenStaff
 import brzzzn.fabadditions.item.warp.network.RequestTeleport
-import brzzzn.fabadditions.screens.FabAdditionsUiScreen
+import brzzzn.fabadditions.ui.screens.FabAdditionsUiScreen
 import com.google.gson.Gson
 import net.fabricmc.api.EnvType
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking
