+++ src/main/kotlin/brzzzn/fabadditions/ui/screens/guis/phantomstaff/PhantomStaffGui.kt
@@ -1,4 +1,4 @@
-package brzzzn.fabadditions.guis.phantomstaff
+package brzzzn.fabadditions.ui.screens.guis.phantomstaff
 
 import brzzzn.fabadditions.data.PlayerList
 import brzzzn.fabadditions.data.PlayerRef
